# docs/swagger.yaml
basePath: /
definitions:
  cart.CartItemParams:
    properties:
      cart_id:
        type: integer
      item_id:
        type: integer
      quantity:
        type: integer
      restaurant_id:
        type: integer
    type: object
  cart.PlaceOrderRequest:
    properties:
      address:
        type: string
    required:
    - address
    type: object
...
host: localhost:8080
info:
  contact:
    name: Akshay Sharma
    url: github.com/akshay4git/Go-Eats
  description: A food delivery backend system with user management, restaurant operations,
    cart/order processing, delivery management with 2FA and real-time notifications.
  title: GO-EATS API
  version: "1.0"
paths:
  /cart/add:
    post:
      consumes:
      - application/json
      description: Adds a menu item to the authenticated user's cart
      parameters:
      - description: Cart item details
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/cart.CartItemParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/list:
    get:
      description: Returns all items in the authenticated user's cart
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List cart items
      tags:
      - Cart
  /cart/order/new:
    post:
      consumes:
      - application/json
      description: Places a new order from the authenticated user's cart
      parameters:
      - description: Delivery address
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/cart.PlaceOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Place a new order
      tags:
      - Orders
  /cart/orders:
    get:
      description: Returns all orders for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List orders
      tags:
      - Orders
  /cart/orders/{id}:
    get:
      description: Returns all items for a specific order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List order items
      tags:
      - Orders
...
  /cart/remove/{id}:
    delete:
      description: Removes a specific item from the cart by cart item ID
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
...
  /restaurant/menu:
    get:
      description: Returns all menu items, or filtered by restaurant_id if provided
      parameters:
      - description: Restaurant ID (optional)
        in: query
        name: restaurant_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List menu items
      tags:
      - Restaurant
    post:
      consumes:
      - application/json
      description: Adds a new menu item to a restaurant
      parameters:
      - description: Menu item details
        in: body
        name: menuItem
        required: true
        schema:
          properties:
            available:
              type: boolean
            category:
              type: string
            description:
              type: string
            name:
              type: string
            price:
              type: number
            restaurant_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a menu item
      tags:
      - Restaurant
  /restaurant/menu/{restaurant_id}/{menu_id}:
    delete:
      description: Deletes a menu item by restaurant ID and menu ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurant_id
        required: true
        type: integer
      - description: Menu ID
        in: path
        name: menu_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a menu item
      tags:
      - Restaurant
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"